+++ README.md
@@ -3,7 +3,7 @@ LearnWebGPU - Code
 
 This repository contains the reference code base accompanying the [Learn WebGPU](learnwgpu.com) web book.
 
-Branch `step010`: This corresponds to the code at the end of the page [The Adapter](learnwgpu.com/getting-started/the-adapter.html).
+Branch `step015`: This corresponds to the code at the end of the page [The Device](learnwgpu.com/getting-started/the-device.html).
 
 Building
 --------
