+++ main.cpp
@@ -25,6 +25,7 @@
  */
 
 #include "glfw3webgpu.h"
+#include "webgpu-utils.h"
 
 #include <GLFW/glfw3.h>
 
@@ -32,121 +33,8 @@
 
 #include <iostream>
 #include <cassert>
-#include <vector>
 
 #define UNUSED(x) (void)x;
-
-/**
- * Utility function to get a WebGPU adapter, so that
- *     WGPUAdapter adapter = requestAdapter(options);
- * is roughly equivalent to
- *     const adapter = await navigator.gpu.requestAdapter(options);
- */
-WGPUAdapter requestAdapter(WGPUInstance instance, WGPURequestAdapterOptions const * options) {
-	// A simple structure holding the local information shared with the
-	// onAdapterRequestEnded callback.
-	struct UserData {
-		WGPUAdapter adapter = nullptr;
-		bool requestEnded = false;
-	};
-	UserData userData;
-
-	// Callback called by wgpuInstanceRequestAdapter when the request returns
-	// This is a C++ lambda function, but could be any function defined in the
-	// global scope. It must be non-capturing (the brackets [] are empty) so
-	// that it behaves like a regular C function pointer, which is what
-	// wgpuInstanceRequestAdapter expects (WebGPU being a C API). The workaround
-	// is to convey what we want to capture through the pUserData pointer,
-	// provided as the last argument of wgpuInstanceRequestAdapter and received
-	// by the callback as its last argument.
-	auto onAdapterRequestEnded = [](WGPURequestAdapterStatus status, WGPUAdapter adapter, char const * message, void * pUserData) {
-		UserData& userData = *reinterpret_cast<UserData*>(pUserData);
-		if (status == WGPURequestAdapterStatus_Success) {
-			userData.adapter = adapter;
-		} else {
-			std::cout << "Could not get WebGPU adapter: " << message << std::endl;
-		}
-		userData.requestEnded = true;
-	};
-
-	// Call to the WebGPU request adapter procedure
-	wgpuInstanceRequestAdapter(
-		instance /* equivalent of navigator.gpu */,
-		options,
-		onAdapterRequestEnded,
-		(void*)&userData
-	);
-
-	// In theory we should wait until onAdapterReady has been called, which
-	// could take some time (what the 'await' keyword does in the JavaScript
-	// code). In practice, we know that when the wgpuInstanceRequestAdapter()
-	// function returns its callback has been called.
-	assert(userData.requestEnded);
-
-	return userData.adapter;
-}
-
-/**
- * An example of how we can inspect the capabilities of the hardware through
- * the adapter object.
- */
-void inspectAdapter(WGPUAdapter adapter) {
-	std::vector<WGPUFeatureName> features;
-	size_t featureCount = wgpuAdapterEnumerateFeatures(adapter, nullptr);
-	features.resize(featureCount);
-	wgpuAdapterEnumerateFeatures(adapter, features.data());
-
-	std::cout << "Adapter features:" << std::endl;
-	for (auto f : features) {
-		std::cout << " - " << f << std::endl;
-	}
-
-	WGPUSupportedLimits limits = {};
-	limits.nextInChain = nullptr;
-	bool success = wgpuAdapterGetLimits(adapter, &limits);
-	if (success) {
-		std::cout << "Adapter limits:" << std::endl;
-		std::cout << " - maxTextureDimension1D: " << limits.limits.maxTextureDimension1D << std::endl;
-		std::cout << " - maxTextureDimension2D: " << limits.limits.maxTextureDimension2D << std::endl;
-		std::cout << " - maxTextureDimension3D: " << limits.limits.maxTextureDimension3D << std::endl;
-		std::cout << " - maxTextureArrayLayers: " << limits.limits.maxTextureArrayLayers << std::endl;
-		std::cout << " - maxBindGroups: " << limits.limits.maxBindGroups << std::endl;
-		std::cout << " - maxDynamicUniformBuffersPerPipelineLayout: " << limits.limits.maxDynamicUniformBuffersPerPipelineLayout << std::endl;
-		std::cout << " - maxDynamicStorageBuffersPerPipelineLayout: " << limits.limits.maxDynamicStorageBuffersPerPipelineLayout << std::endl;
-		std::cout << " - maxSampledTexturesPerShaderStage: " << limits.limits.maxSampledTexturesPerShaderStage << std::endl;
-		std::cout << " - maxSamplersPerShaderStage: " << limits.limits.maxSamplersPerShaderStage << std::endl;
-		std::cout << " - maxStorageBuffersPerShaderStage: " << limits.limits.maxStorageBuffersPerShaderStage << std::endl;
-		std::cout << " - maxStorageTexturesPerShaderStage: " << limits.limits.maxStorageTexturesPerShaderStage << std::endl;
-		std::cout << " - maxUniformBuffersPerShaderStage: " << limits.limits.maxUniformBuffersPerShaderStage << std::endl;
-		std::cout << " - maxUniformBufferBindingSize: " << limits.limits.maxUniformBufferBindingSize << std::endl;
-		std::cout << " - maxStorageBufferBindingSize: " << limits.limits.maxStorageBufferBindingSize << std::endl;
-		std::cout << " - minUniformBufferOffsetAlignment: " << limits.limits.minUniformBufferOffsetAlignment << std::endl;
-		std::cout << " - minStorageBufferOffsetAlignment: " << limits.limits.minStorageBufferOffsetAlignment << std::endl;
-		std::cout << " - maxVertexBuffers: " << limits.limits.maxVertexBuffers << std::endl;
-		std::cout << " - maxVertexAttributes: " << limits.limits.maxVertexAttributes << std::endl;
-		std::cout << " - maxVertexBufferArrayStride: " << limits.limits.maxVertexBufferArrayStride << std::endl;
-		std::cout << " - maxInterStageShaderComponents: " << limits.limits.maxInterStageShaderComponents << std::endl;
-		std::cout << " - maxComputeWorkgroupStorageSize: " << limits.limits.maxComputeWorkgroupStorageSize << std::endl;
-		std::cout << " - maxComputeInvocationsPerWorkgroup: " << limits.limits.maxComputeInvocationsPerWorkgroup << std::endl;
-		std::cout << " - maxComputeWorkgroupSizeX: " << limits.limits.maxComputeWorkgroupSizeX << std::endl;
-		std::cout << " - maxComputeWorkgroupSizeY: " << limits.limits.maxComputeWorkgroupSizeY << std::endl;
-		std::cout << " - maxComputeWorkgroupSizeZ: " << limits.limits.maxComputeWorkgroupSizeZ << std::endl;
-		std::cout << " - maxComputeWorkgroupsPerDimension: " << limits.limits.maxComputeWorkgroupsPerDimension << std::endl;
-	}
-
-	WGPUAdapterProperties properties = {};
-	properties.nextInChain = nullptr;
-	wgpuAdapterGetProperties(adapter, &properties);
-	std::cout << "Adapter properties:" << std::endl;
-	std::cout << " - vendorID: " << properties.vendorID << std::endl;
-	std::cout << " - deviceID: " << properties.deviceID << std::endl;
-	std::cout << " - name: " << properties.name << std::endl;
-	if (properties.driverDescription) {
-		std::cout << " - driverDescription: " << properties.driverDescription << std::endl;
-	}
-	std::cout << " - adapterType: " << properties.adapterType << std::endl;
-	std::cout << " - backendType: " << properties.backendType << std::endl;
-}
 
 int main (int, char**) {
 	WGPUInstanceDescriptor desc = {};
@@ -162,9 +50,7 @@ int main (int, char**) {
 		return 1;
 	}
 
-	// We ask GLFW not to set up any graphics API, we'll do it manually
 	glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
-
 	GLFWwindow* window = glfwCreateWindow(640, 480, "Learn WebGPU", NULL, NULL);
 	if (!window) {
 		std::cerr << "Could not open window!" << std::endl;
@@ -172,22 +58,38 @@ int main (int, char**) {
 	}
 
 	std::cout << "Requesting adapter..." << std::endl;
-
-	// Utility function provided by glfw3webgpu.h
-	WGPUSurface surface = glfwGetWGPUSurface(instance, window);
-
-	// Adapter options: we need the adapter to draw to the window's surface
 	WGPURequestAdapterOptions adapterOpts = {};
 	adapterOpts.nextInChain = nullptr;
-	adapterOpts.compatibleSurface = surface;
-
-	// Get the adapter, see the comments in the definition of the body of the
-	// requestAdapter function above.
+	adapterOpts.compatibleSurface = glfwGetWGPUSurface(instance, window);
 	WGPUAdapter adapter = requestAdapter(instance, &adapterOpts);
-
 	std::cout << "Got adapter: " << adapter << std::endl;
 
-	inspectAdapter(adapter);
+	std::cout << "Requesting device..." << std::endl;
+
+    WGPUDeviceDescriptor deviceDesc = {};
+	deviceDesc.nextInChain = nullptr;
+	deviceDesc.label = "My Device"; // anything works here, that's your call
+	deviceDesc.requiredFeaturesCount = 0; // we do not require any specific feature
+	deviceDesc.requiredLimits = nullptr; // we do not require any specific limit
+	deviceDesc.defaultQueue.nextInChain = nullptr;
+	deviceDesc.defaultQueue.label = "The default queue";
+	WGPUDevice device = requestDevice(adapter, &deviceDesc);
+
+	std::cout << "Got device: " << device << std::endl;
+
+	// Add a callback that gets executed upon errors in our use of the device
+	auto onDeviceError = [](WGPUErrorType type, char const* message, void* /* pUserData */) {
+		std::cout << "Uncaptured device error: type " << type;
+		if (message) std::cout << " (" << message << ")";
+		std::cout << std::endl;
+	};
+	wgpuDeviceSetUncapturedErrorCallback(device, onDeviceError, nullptr /* pUserData */);
+
+	inspectDevice(device);
+
+	wgpuDeviceSetDeviceLostCallback(device, [](WGPUDeviceLostReason reason, char const * message, void*) {
+		std::cout << "Device lost! Reason: " << reason << ", message: " << message << std::endl;
+	}, nullptr);
 
 	while (!glfwWindowShouldClose(window)) {
 		glfwPollEvents();
