+++ webgpu-utils.cpp
@@ -1,0 +1,208 @@
+/**
+ * This file is part of the "Learn WebGPU for C++" book.
+ *   https://github.com/eliemichel/LearnWebGPU
+ * 
+ * MIT License
+ * Copyright (c) 2022 Elie Michel
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#include "webgpu-utils.h"
+
+#include <iostream>
+#include <vector>
+#include <cassert>
+
+WGPUAdapter requestAdapter(WGPUInstance instance, WGPURequestAdapterOptions const * options) {
+	// A simple structure holding the local information shared with the
+	// onAdapterRequestEnded callback.
+	struct UserData {
+		WGPUAdapter adapter = nullptr;
+		bool requestEnded = false;
+	};
+	UserData userData;
+
+	// Callback called by wgpuInstanceRequestAdapter when the request returns
+	// This is a C++ lambda function, but could be any function defined in the
+	// global scope. It must be non-capturing (the brackets [] are empty) so
+	// that it behaves like a regular C function pointer, which is what
+	// wgpuInstanceRequestAdapter expects (WebGPU being a C API). The workaround
+	// is to convey what we want to capture through the pUserData pointer,
+	// provided as the last argument of wgpuInstanceRequestAdapter and received
+	// by the callback as its last argument.
+	auto onAdapterRequestEnded = [](WGPURequestAdapterStatus status, WGPUAdapter adapter, char const * message, void * pUserData) {
+		UserData& userData = *reinterpret_cast<UserData*>(pUserData);
+		if (status == WGPURequestAdapterStatus_Success) {
+			userData.adapter = adapter;
+		} else {
+			std::cout << "Could not get WebGPU adapter: " << message << std::endl;
+		}
+		userData.requestEnded = true;
+	};
+
+	// Call to the WebGPU request adapter procedure
+	wgpuInstanceRequestAdapter(
+		instance /* equivalent of navigator.gpu */,
+		options,
+		onAdapterRequestEnded,
+		(void*)&userData
+	);
+
+	// In theory we should wait until onAdapterReady has been called, which
+	// could take some time (what the 'await' keyword does in the JavaScript
+	// code). In practice, we know that when the wgpuInstanceRequestAdapter()
+	// function returns its callback has been called.
+	assert(userData.requestEnded);
+
+	return userData.adapter;
+}
+
+// This function is very similar to requestAdapter, see comments above
+WGPUDevice requestDevice(WGPUAdapter adapter, WGPUDeviceDescriptor const * descriptor) {
+	struct UserData {
+		WGPUDevice device = nullptr;
+		bool requestEnded = false;
+	};
+	UserData userData;
+
+	auto onDeviceRequestEnded = [](WGPURequestDeviceStatus status, WGPUDevice device, char const * message, void * pUserData) {
+		UserData& userData = *reinterpret_cast<UserData*>(pUserData);
+		if (status == WGPURequestDeviceStatus_Success) {
+			userData.device = device;
+		} else {
+			std::cout << "Could not get WebGPU adapter: " << message << std::endl;
+		}
+		userData.requestEnded = true;
+	};
+
+	wgpuAdapterRequestDevice(
+		adapter,
+		descriptor,
+		onDeviceRequestEnded,
+		(void*)&userData
+	);
+
+	assert(userData.requestEnded);
+
+	return userData.device;
+}
+
+void inspectAdapter(WGPUAdapter adapter) {
+	std::vector<WGPUFeatureName> features;
+	size_t featureCount = wgpuAdapterEnumerateFeatures(adapter, nullptr);
+	features.resize(featureCount);
+	wgpuAdapterEnumerateFeatures(adapter, features.data());
+
+	std::cout << "Adapter features:" << std::endl;
+	for (auto f : features) {
+		std::cout << " - " << f << std::endl;
+	}
+
+	WGPUSupportedLimits limits = {};
+	limits.nextInChain = nullptr;
+	bool success = wgpuAdapterGetLimits(adapter, &limits);
+	if (success) {
+		std::cout << "Adapter limits:" << std::endl;
+		std::cout << " - maxTextureDimension1D: " << limits.limits.maxTextureDimension1D << std::endl;
+		std::cout << " - maxTextureDimension2D: " << limits.limits.maxTextureDimension2D << std::endl;
+		std::cout << " - maxTextureDimension3D: " << limits.limits.maxTextureDimension3D << std::endl;
+		std::cout << " - maxTextureArrayLayers: " << limits.limits.maxTextureArrayLayers << std::endl;
+		std::cout << " - maxBindGroups: " << limits.limits.maxBindGroups << std::endl;
+		std::cout << " - maxDynamicUniformBuffersPerPipelineLayout: " << limits.limits.maxDynamicUniformBuffersPerPipelineLayout << std::endl;
+		std::cout << " - maxDynamicStorageBuffersPerPipelineLayout: " << limits.limits.maxDynamicStorageBuffersPerPipelineLayout << std::endl;
+		std::cout << " - maxSampledTexturesPerShaderStage: " << limits.limits.maxSampledTexturesPerShaderStage << std::endl;
+		std::cout << " - maxSamplersPerShaderStage: " << limits.limits.maxSamplersPerShaderStage << std::endl;
+		std::cout << " - maxStorageBuffersPerShaderStage: " << limits.limits.maxStorageBuffersPerShaderStage << std::endl;
+		std::cout << " - maxStorageTexturesPerShaderStage: " << limits.limits.maxStorageTexturesPerShaderStage << std::endl;
+		std::cout << " - maxUniformBuffersPerShaderStage: " << limits.limits.maxUniformBuffersPerShaderStage << std::endl;
+		std::cout << " - maxUniformBufferBindingSize: " << limits.limits.maxUniformBufferBindingSize << std::endl;
+		std::cout << " - maxStorageBufferBindingSize: " << limits.limits.maxStorageBufferBindingSize << std::endl;
+		std::cout << " - minUniformBufferOffsetAlignment: " << limits.limits.minUniformBufferOffsetAlignment << std::endl;
+		std::cout << " - minStorageBufferOffsetAlignment: " << limits.limits.minStorageBufferOffsetAlignment << std::endl;
+		std::cout << " - maxVertexBuffers: " << limits.limits.maxVertexBuffers << std::endl;
+		std::cout << " - maxVertexAttributes: " << limits.limits.maxVertexAttributes << std::endl;
+		std::cout << " - maxVertexBufferArrayStride: " << limits.limits.maxVertexBufferArrayStride << std::endl;
+		std::cout << " - maxInterStageShaderComponents: " << limits.limits.maxInterStageShaderComponents << std::endl;
+		std::cout << " - maxComputeWorkgroupStorageSize: " << limits.limits.maxComputeWorkgroupStorageSize << std::endl;
+		std::cout << " - maxComputeInvocationsPerWorkgroup: " << limits.limits.maxComputeInvocationsPerWorkgroup << std::endl;
+		std::cout << " - maxComputeWorkgroupSizeX: " << limits.limits.maxComputeWorkgroupSizeX << std::endl;
+		std::cout << " - maxComputeWorkgroupSizeY: " << limits.limits.maxComputeWorkgroupSizeY << std::endl;
+		std::cout << " - maxComputeWorkgroupSizeZ: " << limits.limits.maxComputeWorkgroupSizeZ << std::endl;
+		std::cout << " - maxComputeWorkgroupsPerDimension: " << limits.limits.maxComputeWorkgroupsPerDimension << std::endl;
+	}
+
+	WGPUAdapterProperties properties = {};
+	properties.nextInChain = nullptr;
+	wgpuAdapterGetProperties(adapter, &properties);
+	std::cout << "Adapter properties:" << std::endl;
+	std::cout << " - vendorID: " << properties.vendorID << std::endl;
+	std::cout << " - deviceID: " << properties.deviceID << std::endl;
+	std::cout << " - name: " << properties.name << std::endl;
+	if (properties.driverDescription) {
+		std::cout << " - driverDescription: " << properties.driverDescription << std::endl;
+	}
+	std::cout << " - adapterType: " << properties.adapterType << std::endl;
+	std::cout << " - backendType: " << properties.backendType << std::endl;
+}
+
+void inspectDevice(WGPUDevice device) {
+	std::vector<WGPUFeatureName> features;
+	size_t featureCount = wgpuDeviceEnumerateFeatures(device, nullptr);
+	features.resize(featureCount);
+	wgpuDeviceEnumerateFeatures(device, features.data());
+
+	std::cout << "Device features:" << std::endl;
+	for (auto f : features) {
+		std::cout << " - " << f << std::endl;
+	}
+
+	WGPUSupportedLimits limits = {};
+	limits.nextInChain = nullptr;
+	bool success = wgpuDeviceGetLimits(device, &limits);
+	if (success) {
+		std::cout << "Device limits:" << std::endl;
+		std::cout << " - maxTextureDimension1D: " << limits.limits.maxTextureDimension1D << std::endl;
+		std::cout << " - maxTextureDimension2D: " << limits.limits.maxTextureDimension2D << std::endl;
+		std::cout << " - maxTextureDimension3D: " << limits.limits.maxTextureDimension3D << std::endl;
+		std::cout << " - maxTextureArrayLayers: " << limits.limits.maxTextureArrayLayers << std::endl;
+		std::cout << " - maxBindGroups: " << limits.limits.maxBindGroups << std::endl;
+		std::cout << " - maxDynamicUniformBuffersPerPipelineLayout: " << limits.limits.maxDynamicUniformBuffersPerPipelineLayout << std::endl;
+		std::cout << " - maxDynamicStorageBuffersPerPipelineLayout: " << limits.limits.maxDynamicStorageBuffersPerPipelineLayout << std::endl;
+		std::cout << " - maxSampledTexturesPerShaderStage: " << limits.limits.maxSampledTexturesPerShaderStage << std::endl;
+		std::cout << " - maxSamplersPerShaderStage: " << limits.limits.maxSamplersPerShaderStage << std::endl;
+		std::cout << " - maxStorageBuffersPerShaderStage: " << limits.limits.maxStorageBuffersPerShaderStage << std::endl;
+		std::cout << " - maxStorageTexturesPerShaderStage: " << limits.limits.maxStorageTexturesPerShaderStage << std::endl;
+		std::cout << " - maxUniformBuffersPerShaderStage: " << limits.limits.maxUniformBuffersPerShaderStage << std::endl;
+		std::cout << " - maxUniformBufferBindingSize: " << limits.limits.maxUniformBufferBindingSize << std::endl;
+		std::cout << " - maxStorageBufferBindingSize: " << limits.limits.maxStorageBufferBindingSize << std::endl;
+		std::cout << " - minUniformBufferOffsetAlignment: " << limits.limits.minUniformBufferOffsetAlignment << std::endl;
+		std::cout << " - minStorageBufferOffsetAlignment: " << limits.limits.minStorageBufferOffsetAlignment << std::endl;
+		std::cout << " - maxVertexBuffers: " << limits.limits.maxVertexBuffers << std::endl;
+		std::cout << " - maxVertexAttributes: " << limits.limits.maxVertexAttributes << std::endl;
+		std::cout << " - maxVertexBufferArrayStride: " << limits.limits.maxVertexBufferArrayStride << std::endl;
+		std::cout << " - maxInterStageShaderComponents: " << limits.limits.maxInterStageShaderComponents << std::endl;
+		std::cout << " - maxComputeWorkgroupStorageSize: " << limits.limits.maxComputeWorkgroupStorageSize << std::endl;
+		std::cout << " - maxComputeInvocationsPerWorkgroup: " << limits.limits.maxComputeInvocationsPerWorkgroup << std::endl;
+		std::cout << " - maxComputeWorkgroupSizeX: " << limits.limits.maxComputeWorkgroupSizeX << std::endl;
+		std::cout << " - maxComputeWorkgroupSizeY: " << limits.limits.maxComputeWorkgroupSizeY << std::endl;
+		std::cout << " - maxComputeWorkgroupSizeZ: " << limits.limits.maxComputeWorkgroupSizeZ << std::endl;
+		std::cout << " - maxComputeWorkgroupsPerDimension: " << limits.limits.maxComputeWorkgroupsPerDimension << std::endl;
+	}
+}
